Data entry

+++ banking-service/src/main/java/com/deloittevg/controller/BankController.java
@@ -2,7 +2,6 @@ package com.deloittevg.controller;
 
 import com.deloittevg.entity.BankAccount;
 import com.deloittevg.service.BankAccountService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,9 +12,13 @@ import java.util.List;
 @RequestMapping("/bank")
 public class BankController {
 	
-	@Autowired
-	BankAccountService bankAccountService;
-	
+
+	private final BankAccountService bankAccountService;
+
+	public BankController(BankAccountService bankAccountService) {
+		this.bankAccountService = bankAccountService;
+	}
+
 	@GetMapping("/accounts")
 	public ResponseEntity<List<BankAccount>> viewAll(){
 		List<BankAccount>  account  = bankAccountService.viewAllAccounts();
@@ -38,34 +41,46 @@ public class BankController {
 			
 		}
 
-	@GetMapping ("/accounts/{userId}")
+	@GetMapping ("/accounts/user/{userId}")
 	public ResponseEntity<List<BankAccount>> searchByUser(@PathVariable long userId) {
 		List<BankAccount> account1 = bankAccountService.searchByUserId(userId);
 		if(account1.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+			return ResponseEntity.status(HttpStatus.OK).build();
 		}
 		return ResponseEntity.status(HttpStatus.OK).body(account1);
 
 	}
 
 	@PutMapping("/accounts/{accountNo}/update")
-	public ResponseEntity<BankAccount> updateAccount(@RequestBody BankAccount account,@PathVariable String accountNo) {
-		account.setAccountNo(accountNo);
-		BankAccount account1= bankAccountService.updateAccount(account);
-		if(account1==null) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+	public ResponseEntity<String> updateAccount(@RequestBody BankAccount account,@PathVariable String accountNo) {
+
+		try {
+			account.setAccountNo(accountNo);
+			BankAccount account1 = bankAccountService.findByAccountNo(accountNo);
+			if (account1 == null) {
+				return new ResponseEntity<>("Account not found", HttpStatus.NOT_FOUND);
+			} else {
+				account1.setAccountNo(accountNo);
+				bankAccountService.updateAccount(account);
+				return new ResponseEntity<>("Account updated successfully !", HttpStatus.OK);
+			}
+		} catch (Exception e) {
+			return new ResponseEntity<>("Error, Cannot Update Account !" + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
 		}
-			return ResponseEntity.status(HttpStatus.OK).body(account1);
-			
-		}
-	
-	
-	
+	}
+
+
 	@DeleteMapping("/accounts/{accountNo}/delete")
 	public ResponseEntity<String> deleteAccount(@PathVariable String accountNo){
 		try {
-			bankAccountService.deleteAccount(accountNo);
-			return new ResponseEntity<>("Account Deleted Successfully !", HttpStatus.OK);
+			BankAccount account =  bankAccountService.findByAccountNo(accountNo);
+			if(account == null){
+				return new ResponseEntity<>("Account not found", HttpStatus.NOT_FOUND);
+			}
+			else{
+				bankAccountService.deleteAccount(accountNo);
+				return new ResponseEntity<>("Account Deleted Successfully !", HttpStatus.OK);
+			}
 		}
 		catch (Exception e) {
 			return new ResponseEntity<>("Error, Cannot Delete Account !" + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
